SQLdata file
This is the SQL data file on which the VBA code is run for Forecasting using Holt's Winter Multiplicative Time Series Decomposition. Click continue if 'the number of feasible solutions has reached' dialog box appears. This appears when the optimization reaches the minimum value for U-Statistic. It will output two other tabs Forecast and the Trend.